+++ composer.json
@@ -11,7 +11,9 @@
     "ext-gd": "*"
   },
   "require-dev": {
-    "phpunit/phpunit": "5.7.*|7.4.*|^9.6",
+    "php": ">=8.1",
+    "phpunit/phpunit": "^9.6",
+    "meyfa/php-svg": "^v0.15",
     "ext-imagick": "*"
   }
 }
